requirements.txt
absl_py==1.0.0
absl_py==1.4.0
edward2==0.0.2
keras==2.12.0
keras_nightly==2.11.0.dev2022101507
matplotlib==3.3.4
numpy==1.23.5
robustness_metrics==0.0.1
scikit_learn==1.1.2
tb_nightly==2.11.0a20221015
tensorboard==2.12.3
tensorflow==2.12.0
tensorflow_datasets==4.9.2
tensorflow_probability==0.20.0
tf_nightly==2.11.0.dev20221015
tfds_nightly==4.4.0.dev202111160106
tensorflow_gpu==2.9.0
tensorflow-addons==0.13.0
